add video module
