+++ apps/api-journeys/src/modules/block/block.spec.ts
@@ -319,44 +319,4 @@ describe('BlockModule', () => {
       }
     ])
   })
-
-  it('returns VideoBlock', async () => {
-    const parentBlockId = uuidv4()
-    const video: Block = {
-      id: uuidv4(),
-      journeyId,
-      blockType: 'VideoBlock',
-      parentBlockId,
-      parentOrder: 1,
-      extraAttrs: {
-        src: 'src',
-        title: 'title'
-      }
-    }
-    dbMock.block.findMany.mockResolvedValue([video])
-    const { data } = await query(gql`
-      query ($id: ID!) {
-        journey(id: $id) {
-          blocks {
-            id
-            __typename
-            parentBlockId
-            ... on VideoBlock {
-              src
-              title
-            }
-          }
-        }
-      }
-    `)
-    expect(data?.journey.blocks).toEqual([
-      {
-        id: video.id,
-        __typename: 'VideoBlock',
-        parentBlockId,
-        src: 'src',
-        title: 'title'
-      }
-    ])
-  })
 })
