+++ apps/api-journeys/src/modules/index.ts
@@ -4,3 +4,4 @@ export { journeyModule } from './journey'
 export { radioQuestionModule } from './radioQuestion'
 export { responseModule } from './response'
 export { signUpModule } from './signUp'
+export { videoModule } from './video'
