+++ apps/api-journeys/src/modules/radioQuestion/radioQuestion.spec.ts
@@ -1,7 +1,12 @@
 import { testkit, gql } from 'graphql-modules'
 import { schemaBuilder } from '@core/shared/util-graphql'
-import { radioQuestionModule } from '.'
-import { actionModule, journeyModule, blockModule, responseModule } from '..'
+import {
+  radioQuestionModule,
+  actionModule,
+  journeyModule,
+  blockModule,
+  responseModule
+} from '..'
 import dbMock from '../../../tests/dbMock'
 import { v4 as uuidv4 } from 'uuid'
 import {
